fix(compose): Circular shared chat badges, username click behavior setting, and centered tab scrolling

+++ app/src/main/kotlin/com/flxrs/dankchat/ui/chat/messages/common/MessageTextRenderer.kt
@@ -38,7 +38,7 @@ fun MessageTextWithInlineContent(
     onTextClick: (Int) -> Unit,
     onEmoteClick: (List<ChatMessageEmote>) -> Unit,
     modifier: Modifier = Modifier,
-    onTextLongClick: (() -> Unit)? = null,
+    onTextLongClick: ((Int) -> Unit)? = null,
     interactionSource: MutableInteractionSource? = null,
 ) {
     val emoteCoordinator = LocalEmoteAnimationCoordinator.current
